+++ frame/undercollateralized-loans/src/helpers/general.rs
@@ -1,7 +1,7 @@
 use crate::{
 	types::{
-		LoanConfigOf, LoanInputOf, MarketConfigOf, MarketInfoOf, MarketInputOf, Payment, PaymentOf,
-		PaymentOutcome, PaymentOutcomeOf, PaymentsOutcomes, Timestamp,
+		LoanConfigOf, LoanInputOf, MarketInfoOf, MarketInputOf, Payment, PaymentOf, PaymentOutcome,
+		PaymentOutcomeOf, PaymentsOutcomes, Timestamp,
 	},
 	validation::{AssetIsSupportedByOracle, CurrencyPairIsNotSame, LoanInputIsValid},
 	Config, DebtTokenForMarketStorage, Error, MarketsStorage, Pallet,
@@ -13,27 +13,27 @@ use composable_traits::{
 	liquidation::Liquidation,
 	oracle::Oracle,
 	undercollateralized_loans::{LoanConfig, MarketConfig, MarketInfo, UndercollateralizedLoans},
-	vault::{Deposit, FundsAvailability, StrategicVault, Vault, VaultConfig},
+	vault::{Deposit, Vault, VaultConfig},
 };
 use frame_support::{
 	ensure,
-	storage::with_transaction,
 	traits::{
 		fungible::{Inspect as NativeInspect, Transfer as NativeTransfer},
-		fungibles::{Inspect, Mutate, Transfer},
+		fungibles::{Mutate, Transfer},
 		Get, UnixTime,
 	},
 };
-
 use sp_runtime::{
 	traits::{One, Saturating, Zero},
-	DispatchError, Perquintill, TransactionOutcome,
+	DispatchError, Perquintill,
 };
-
-use chrono::{Duration, NaiveDate, NaiveDateTime, NaiveTime};
-use sp_std::{ops::Add, vec::Vec};
+use sp_std::vec::Vec;
 
 impl<T: Config> Pallet<T> {
+	//===========================================================================================================================
+	//                                                   "Do" methods
+	//===========================================================================================================================
+
 	pub(crate) fn do_create_market(
 		manager: T::AccountId,
 		input: Validated<
@@ -94,8 +94,8 @@ impl<T: Config> Pallet<T> {
 			Ok(market_info)
 		})
 	}
+
 	// Create non-active loan, which should be activated via borrower.
-	// TODO: @mikolaichuk: check why LoanInputOf does not work here
 	pub(crate) fn do_create_loan(
 		input: Validated<LoanInputOf<T>, LoanInputIsValid<crate::Pallet<T>>>,
 	) -> Result<LoanConfigOf<T>, DispatchError> {
@@ -104,7 +104,7 @@ impl<T: Config> Pallet<T> {
 		// validation process.
 		let market_config =
 			Self::get_market_config_via_account_id(&config_input.market_account_id)?;
-		// Align schedule timstamps to the beginning of the day.
+		// Align schedule timestamps to the beginning of the day.
 		// 24.08.1991 08:45:03 -> 24.08.1991 00:00:00
 		let schedule = config_input
 			.payment_schedule
@@ -228,53 +228,32 @@ impl<T: Config> Pallet<T> {
 					Self::process_checked_payment(&payment),
 				PaymentOutcome::LastPaymentSucceed(payment) => {
 					Self::process_checked_payment(&payment);
-					Self::do_close_loan_contract(&payment.loan_config);
+					Self::close_loan_contract(&payment.loan_config);
 				},
 				PaymentOutcome::PaymentFailed(payment) =>
-					Self::do_process_failed_payment(&payment.loan_config),
+					Self::process_failed_payment(&payment.loan_config),
 			}
 		}
 	}
 
-	// Process payments which for some reason were not processed EOD.
-	pub(crate) fn do_process_unchecked_payments(
-		loans_accounts_ids: Vec<T::AccountId>,
-		date: Timestamp,
-	) {
-		for loan_account_id in loans_accounts_ids {
-			let loan_config = match Self::get_loan_config_via_account_id(&loan_account_id) {
-				Ok(loan_config) => loan_config,
-				Err(error) => {
-					log::error!("Error: {:?}", error);
-					continue
-				},
-			};
-			let payment = match Self::treat_payment(&loan_config, date) {
-				Some(payment) => payment,
-				None => continue,
-			};
-			match payment {
-				PaymentOutcome::RegularPaymentSucceed(_) => (),
-				PaymentOutcome::LastPaymentSucceed(payment) =>
-					Self::do_close_loan_contract(&payment.loan_config),
-				PaymentOutcome::PaymentFailed(payment) =>
-					Self::do_process_failed_payment(&payment.loan_config),
-			}
-		}
+	//Removes a bunch of non-activated loans.
+	//TODO: @mikolaichuk: add Event;
+	pub(crate) fn do_remove_non_activated_loans(loans_accounts_ids: Vec<T::AccountId>) {
+		loans_accounts_ids
+			.into_iter()
+			//Check that loan is non-activated.
+			.filter(|loan_account_id| {
+				crate::NonActiveLoansStorage::<T>::contains_key(loan_account_id)
+			})
+			.for_each(|loan_account_id| {
+				crate::LoansStorage::<T>::remove(loan_account_id.clone());
+				crate::NonActiveLoansStorage::<T>::remove(loan_account_id)
+			});
 	}
 
-	// Collects accounts ids of unprocessed loans and process them.
-	// Used at the beginning of the day to proccess loans which
-	// were not processed yesterday for some reason.
-	pub(crate) fn last_chance_processing(date: Timestamp) {
-		let unprocessed_loans_accounts_ids: Vec<_> = crate::ScheduleStorage::<T>::get(date)
-			.difference(&crate::ProcessedLoansStorage::<T>::get())
-			.cloned()
-			.collect();
-		if !unprocessed_loans_accounts_ids.is_empty() {
-			Self::do_process_unchecked_payments(unprocessed_loans_accounts_ids, date);
-		};
-	}
+	//===========================================================================================================================
+	//                                          General payments treatment
+	//===========================================================================================================================
 
 	// In off-chain context: checks that payment will be successful.
 	// In on-chain context: really transfers assets from borrower's account to
@@ -287,7 +266,8 @@ impl<T: Config> Pallet<T> {
 		let payment_amount = loan_config.get_payment_for_particular_moment(&today)?;
 		let payment =
 			Payment { loan_config: loan_config.clone(), amount: payment_amount, timestamp: today };
-		// Use to check if payment transfer is possible.
+		// Use to check if payment transfer is possible if off-chain context
+		// In on-chain context try to transfer borrow asset from loan's account to market's account.
 		// Please note that methods called within off-chain context do not change chain's state.
 		let outcome = match Self::pay_back_borrowed_asset(&loan_config, payment_amount) {
 			// We have enough money on the loan's account to perform last payment.
@@ -303,7 +283,8 @@ impl<T: Config> Pallet<T> {
 		Some(outcome)
 	}
 
-	//
+	// Process loans which payments were checked off-chain.
+	// We may be sure that loans' accounts have enough money to pay.
 	pub fn process_checked_payment(payment: &PaymentOf<T>) {
 		let loan_account_id = payment.loan_config.account_id();
 		// Check if the loan was already processed.
@@ -317,25 +298,44 @@ impl<T: Config> Pallet<T> {
 			Some(amount) => amount,
 			None => {
 				log::error!(
-                    "Payment moment does not exist in the loan's local payment schedule, loan account id: {:?}, moment: {:?}.",
-                    loan_account_id, 
-                    payment.timestamp
-                );
+					"Payment moment does not exist in the loan's local payment schedule, loan account id: {:?}, moment: {:?}.",
+					loan_account_id,
+					payment.timestamp
+				);
 				return
 			},
 		};
 		// We are sure that payment is succeed since it has been checked off-chain.
-		// Nobody except the pallet can withdraw money from the loan's account.
+		// Nobody except the pallet can withdraw money from the loan's account, so there is no
+		// chance that somebody withdraw money from loan's account after it was checked off-chain.
 		Self::pay_back_borrowed_asset(&payment.loan_config, amount).unwrap_or(T::Balance::zero());
 		// Mark processed loan to avoid double processing which can cause loan liquidation.
 		crate::ProcessedLoansStorage::<T>::mutate(|set| set.insert(loan_account_id.clone()));
 	}
 
+	// Tansfers borrow asset from loan's account to market's account.
+	pub(crate) fn pay_back_borrowed_asset(
+		loan_config: &LoanConfigOf<T>,
+		payment_amount: T::Balance,
+	) -> Result<T::Balance, DispatchError> {
+		T::MultiCurrency::transfer(
+			*loan_config.borrow_asset_id(),
+			loan_config.account_id(),
+			loan_config.market_account_id(),
+			payment_amount,
+			true,
+		)
+	}
+
+	//===========================================================================================================================
+	//                                          Failed payments treatment
+	//===========================================================================================================================
+
 	// Logic we apply if payment is failed.
-	pub(crate) fn do_process_failed_payment(loan_config: &LoanConfigOf<T>) {
+	pub(crate) fn process_failed_payment(loan_config: &LoanConfigOf<T>) {
 		// TODO: @mikolaichuk:  add credit rating functionality.
 		//                      add event deposit(?)
-		match Self::do_liquidate(loan_config) {
+		match Self::liquidate(loan_config) {
 			Ok(_) => log::info!(
 				"Loan with the following account id: {:?} was successfuly send to liquidation",
 				loan_config.account_id()
@@ -349,7 +349,7 @@ impl<T: Config> Pallet<T> {
 	}
 
 	// Send position to liquidation.
-	pub(crate) fn do_liquidate(
+	pub(crate) fn liquidate(
 		loan_config: &LoanConfigOf<T>,
 	) -> Result<<T::Liquidation as Liquidation>::OrderId, DispatchError> {
 		let liquidation_strategies =
@@ -368,8 +368,12 @@ impl<T: Config> Pallet<T> {
 		T::Liquidation::liquidate(loan_config.account_id(), sell, liquidation_strategies)
 	}
 
+	//===========================================================================================================================
+	//                                              Close contracts procedure
+	//===========================================================================================================================
+
 	// Close loan contract since loan is paid.
-	pub(crate) fn do_close_loan_contract(loan_config: &LoanConfigOf<T>) {
+	pub(crate) fn close_loan_contract(loan_config: &LoanConfigOf<T>) {
 		// Transfer collateral to borrower's account.
 		T::MultiCurrency::transfer(
 			*loan_config.collateral_asset_id(),
@@ -387,147 +391,6 @@ impl<T: Config> Pallet<T> {
 		// Remove all information about the loan.
 		Self::terminate_activated_loan(&loan_config);
 		Self::deposit_event(crate::Event::<T>::LoanClosed { loan_config: loan_config.clone() });
-	}
-
-	// Check if borrower's account is whitelisted for particular market.
-	pub(crate) fn is_borrower_account_whitelisted(
-		borrower_account_id_ref: &T::AccountId,
-		market_account_id_ref: &T::AccountId,
-	) -> Result<bool, DispatchError> {
-		let market_config = Self::get_market_config_via_account_id(market_account_id_ref)?;
-		Ok(market_config.whitelist().contains(borrower_account_id_ref))
-	}
-
-	pub(crate) fn calculate_initial_pool_size(
-		borrow_asset: <T::Oracle as composable_traits::oracle::Oracle>::AssetId,
-	) -> Result<<T as DeFiComposableConfig>::Balance, DispatchError> {
-		T::Oracle::get_price_inverse(borrow_asset, T::OracleMarketCreationStake::get())
-	}
-
-	// Check if provided account id belongs to the market manager.
-	pub(crate) fn is_market_manager_account(
-		account_id: &T::AccountId,
-		market_account_id_ref: &T::AccountId,
-	) -> Result<bool, DispatchError> {
-		let market_config = Self::get_market_config_via_account_id(market_account_id_ref)?;
-		Ok(market_config.manager() == account_id)
-	}
-
-	// TODO: @mikolaichuk: Add weights calculation
-	// #generalization
-	pub(crate) fn treat_vaults_balance(block_number: T::BlockNumber) -> () {
-		let _ = with_transaction(|| {
-			let mut errors = crate::MarketsStorage::<T>::iter()
-				.map(|(market_account_id, market_info)| {
-					match Self::available_funds(&market_info.config(), &market_account_id)? {
-						FundsAvailability::Withdrawable(balance) => {
-							Self::handle_withdrawable(
-								&market_info.config(),
-								&market_account_id,
-								balance,
-							)?;
-						},
-						FundsAvailability::Depositable(balance) => {
-							Self::handle_depositable(
-								&market_info.config(),
-								&market_account_id,
-								balance,
-							)?;
-						},
-						FundsAvailability::MustLiquidate => {
-							Self::handle_must_liquidate(&market_info.config(), market_account_id)?;
-						},
-					}
-					Result::<(), DispatchError>::Ok(())
-				})
-				.filter_map(|r| match r {
-					Ok(_) => None,
-					Err(err) => Some(err),
-				})
-				.peekable();
-
-			if errors.peek().is_none() {
-				TransactionOutcome::Commit(Ok(1000))
-			} else {
-				errors.for_each(|e| {
-					log::error!(
-						"This should never happen, could not initialize block!!! {:#?} {:#?}",
-						block_number,
-						e
-					)
-				});
-				TransactionOutcome::Rollback(Err(DispatchError::Other(
-					"failed to initialize block",
-				)))
-			}
-		});
-	}
-
-	// Check if vault balanced or we have to deposit money to the vault or withdraw money from it.
-	// If vault is balanced we will do nothing.
-	// #generalization
-	fn available_funds(
-		config: &MarketConfigOf<T>,
-		market_account: &T::AccountId,
-	) -> Result<FundsAvailability<T::Balance>, DispatchError> {
-		<T::Vault as StrategicVault>::available_funds(&config.borrow_asset_vault(), market_account)
-	}
-
-	// If we can withdraw from the vault
-	// #generalization
-	fn handle_withdrawable(
-		config: &MarketConfigOf<T>,
-		market_account: &T::AccountId,
-		balance: T::Balance,
-	) -> Result<(), DispatchError> {
-		<T::Vault as StrategicVault>::withdraw(
-			&config.borrow_asset_vault(),
-			market_account,
-			balance,
-		)
-	}
-
-	// If vault is unblanced and we have to deposit some assets to the vault.
-	// #generalization
-	fn handle_depositable(
-		config: &MarketConfigOf<T>,
-		market_account: &T::AccountId,
-		balance: T::Balance,
-	) -> Result<(), DispatchError> {
-		let asset_id = <T::Vault as Vault>::asset_id(&config.borrow_asset_vault())?;
-		let balance =
-			<T as Config>::MultiCurrency::reducible_balance(asset_id, market_account, false)
-				.min(balance);
-		<T::Vault as StrategicVault>::deposit(&config.borrow_asset_vault(), market_account, balance)
-	}
-
-	// TODO: @mikolaichuk: Implement logic when vault is stopped, tombstoned or does not exist.
-	// #generalization
-	fn handle_must_liquidate(
-		_market_config: &MarketConfigOf<T>,
-		_market_account_id: T::AccountId,
-	) -> Result<(), DispatchError> {
-		todo!()
-	}
-
-	// #generalization
-	pub(crate) fn now() -> Timestamp {
-		T::UnixTime::now().as_secs() as Timestamp
-	}
-
-	//Removes a bunch of non-activated loans.
-	//TODO: @mikolaichuk: add Event;
-	pub(crate) fn do_remove_non_activated_loans(loans_accounts_ids: Vec<T::AccountId>) {
-		loans_accounts_ids
-			.into_iter()
-			//Check that loan is non-activated.
-			.filter(|loan_account_id| {
-				crate::NonActiveLoansStorage::<T>::contains_key(loan_account_id)
-			})
-			.for_each(|loan_account_id| {
-				crate::LoansStorage::<T>::remove(loan_account_id.clone());
-				crate::NonActiveLoansStorage::<T>::remove(loan_account_id)
-			});
 	}
 
 	// Remove all information regarding activated loan.
@@ -558,71 +421,36 @@ impl<T: Config> Pallet<T> {
 		Self::deposit_event(crate::Event::<T>::LoanTerminated { loan_config: loan_config.clone() })
 	}
 
-	// Tansfers borrow asset from loan's account to market's account.
-	pub(crate) fn pay_back_borrowed_asset(
-		loan_config: &LoanConfigOf<T>,
-		payment_amount: T::Balance,
-	) -> Result<T::Balance, DispatchError> {
-		T::MultiCurrency::transfer(
-			*loan_config.borrow_asset_id(),
-			loan_config.account_id(),
-			loan_config.market_account_id(),
-			payment_amount,
-			true,
-		)
-	}
+	//===========================================================================================================================
+	//                                              Non-organized helpers
+	//===========================================================================================================================
 
-	// Getters helpers
-	pub(crate) fn get_market_info_via_account_id(
+	// Check if borrower's account is whitelisted for particular market.
+	pub(crate) fn is_borrower_account_whitelisted(
+		borrower_account_id_ref: &T::AccountId,
 		market_account_id_ref: &T::AccountId,
-	) -> Result<MarketInfoOf<T>, crate::Error<T>> {
-		crate::MarketsStorage::<T>::try_get(market_account_id_ref)
-			.map_err(|_| crate::Error::<T>::MarketDoesNotExist)
+	) -> Result<bool, DispatchError> {
+		let market_config = Self::get_market_config_via_account_id(market_account_id_ref)?;
+		Ok(market_config.whitelist().contains(borrower_account_id_ref))
 	}
 
-	pub(crate) fn get_market_config_via_account_id(
+	// Check if provided account id belongs to the market manager.
+	pub(crate) fn is_market_manager_account(
+		account_id: &T::AccountId,
 		market_account_id_ref: &T::AccountId,
-	) -> Result<MarketConfigOf<T>, crate::Error<T>> {
-		let market_info = Self::get_market_info_via_account_id(market_account_id_ref)?;
-		Ok(market_info.config().clone())
+	) -> Result<bool, DispatchError> {
+		let market_config = Self::get_market_config_via_account_id(market_account_id_ref)?;
+		Ok(market_config.manager() == account_id)
 	}
 
-	pub(crate) fn get_loan_config_via_account_id(
-		loan_account_id_ref: &T::AccountId,
-	) -> Result<LoanConfigOf<T>, crate::Error<T>> {
-		crate::LoansStorage::<T>::try_get(loan_account_id_ref)
-			.map_err(|_| crate::Error::<T>::LoanNotFound)
+	pub(crate) fn calculate_initial_pool_size(
+		borrow_asset: <T::Oracle as composable_traits::oracle::Oracle>::AssetId,
+	) -> Result<<T as DeFiComposableConfig>::Balance, DispatchError> {
+		T::Oracle::get_price_inverse(borrow_asset, T::OracleMarketCreationStake::get())
 	}
 
-	pub(crate) fn get_current_date_timestamp() -> Timestamp {
-		crate::CurrentDateStorage::<T>::get()
-	}
-	// Get current date from the storage.
-	pub(crate) fn get_current_date() -> NaiveDate {
-		Self::get_date_from_timestamp(Self::get_current_date_timestamp())
-	}
-	// Get date from a timestamp
-	pub(crate) fn get_date_from_timestamp(timestamp: Timestamp) -> NaiveDate {
-		NaiveDateTime::from_timestamp(timestamp, 0).date()
-	}
-
-	// Align a timestamp to the beginign of the day.
-	// 24.08.1991 08:45:03 -> 24.08.1991 00:00:00
-	// (in terms of seconds from the beginning of Unix epoche)
-	pub(crate) fn get_date_aligned_timestamp(timestamp: Timestamp) -> Timestamp {
-		Self::get_date_from_timestamp(timestamp)
-			.and_time(NaiveTime::default())
-			.timestamp()
-	}
-
-	// Returns next date aligned timestamp.
-	// 24.08.1991 08:45:03 -> 25.08.1991 00:00:00
-	// (in terms of seconds from the beginning of Unix epoche)
-	pub(crate) fn get_next_date_aligned_timestamp(timestamp: Timestamp) -> Timestamp {
-		Self::get_date_from_timestamp(timestamp)
-			//	Gonna no overflow since we adds only one day.
-			.add(Duration::days(1))
-			.and_time(NaiveTime::default())
-			.timestamp()
+	// #generalization
+	pub(crate) fn now() -> Timestamp {
+		T::UnixTime::now().as_secs() as Timestamp
 	}
 }
