+++ frame/undercollateralized-loans/src/helpers/mod.rs
@@ -1,2 +1,8 @@
+/// Helpers which are mostly used to process payments.
 pub mod general;
+/// Getters helpers.
+pub mod getters;
+/// Off-chain context helpers.
 pub mod offchain;
+/// Helpers which are used for block initialization.
+pub mod on_initialize;
